Network client extension moved to general toolkit.
